requirements.txt
flask==3.1.1
flask-cors==6.0.1

langchain==0.1.16
langchain-openai==0.0.8
langchain_community==0.0.32
openai==1.21.0

faiss-cpu==1.8.0.post1

python-dotenv==1.1.1
pydantic==1.10.12
numpy==1.24.4
httpx<0.28